requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# TGMS-Bot\cogs\moderation.py: 5,6
colored == 1.4.2

# TGMS-Bot\bot.py: 2,5
# TGMS-Bot\cogs\moderation.py: 2,3
discord.py == 1.7.2
